#1125 Move processing sample to correct package
