+++ hartshorn-launchpad-samples/application-processing/src/main/java/org/dockbox/sample/processing/HelloWorldSupplier.java
@@ -1,4 +1,4 @@
-package org.dockbox.sample.java;
+package org.dockbox.sample.processing;
 
 public interface HelloWorldSupplier {
