+++ hartshorn-launchpad-samples/application-processing/src/main/java/org/dockbox/sample/processing/UseCustomProcessors.java
@@ -1,13 +1,13 @@
-package org.dockbox.sample.java;
+package org.dockbox.sample.processing;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 import org.dockbox.hartshorn.launchpad.activation.ServiceActivator;
-import org.dockbox.sample.java.bindings.CustomBindingPostProcessor;
-import org.dockbox.sample.java.components.CustomComponentPostProcessor;
-import org.dockbox.sample.java.components.CustomComponentPreProcessor;
+import org.dockbox.sample.processing.bindings.CustomBindingPostProcessor;
+import org.dockbox.sample.processing.components.CustomComponentPostProcessor;
+import org.dockbox.sample.processing.components.CustomComponentPreProcessor;
 
 @ServiceActivator(
     binderPostProcessors = CustomBindingPostProcessor.class,
